% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erf.R
\name{erf}
\alias{erf}
\title{erf activation function}
\usage{
erf(x)
}
\arguments{
\item{x}{a real number}
}
\value{
\code{erf(x)}, erf value
}
\description{
erf activation function
}
\examples{
erf(10)
}
\keyword{methods}
\keyword{models}
