% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.cpen.R
\name{estimate.gamma.cpen_2}
\alias{estimate.gamma.cpen_2}
\title{Estimate marginal log posterior of a single BGNLM model with alternative defaults}
\usage{
estimate.gamma.cpen_2(
  formula,
  data,
  r = 1/223,
  logn = log(223),
  relat = c("to23", "expi", "logi", "to35", "sini", "troot", "sigmoid")
)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{r}{prior inclusion penalty parameter}

\item{logn}{logn}

\item{relat}{a set of nonlinear transformations in the class of BGNLMs of interest}
}
\value{
A list of
\describe{
 \item{mlik}{marginal likelihood of the model}
 \item{waic}{AIC model selection criterion}
 \item{dic}{BIC model selection criterion}
 \item{summary.fixed$mean}{a vector of posterior modes of the parameters}
}
}
\description{
Estimate marginal log posterior of a single BGNLM model with alternative defaults
}
