% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factorial}
\alias{factorial}
\title{Truncated factorial to avoid stack overflow for huge values}
\usage{
factorial(x)
}
\arguments{
\item{x}{a non-negative integer number}
}
\value{
\code{factorial(x)}, truncated facctorial as min(x!,171!)
}
\description{
truncated factorial to avoid stack overflow for huge values
}
\examples{
factorial(10)
}
\keyword{methods}
\keyword{models}
