% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.R
\name{m}
\alias{m}
\title{Product function used in the deep regression context}
\usage{
m(a, b)
}
\arguments{
\item{a}{the first argument}

\item{b}{the second argument}
}
\value{
\code{m(a,b)}, product of the arguments a*b
}
\description{
Product function used in the deep regression context
}
\examples{
m(10,2)
}
\keyword{methods}
\keyword{models}
