% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid activation function}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{a real number}
}
\value{
sigmoid value
}
\description{
sigmoid activation function
}
\examples{
sigmoid(10)
}
\keyword{methods}
\keyword{models}
