% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{truncfactorial}
\alias{truncfactorial}
\title{Truncated factorial to avoid stack overflow for huge values}
\usage{
truncfactorial(x)
}
\arguments{
\item{x}{a non-negative integer number}
}
\value{
\code{truncfactorial(x)}, truncated facctorial as min(x!,171!)
}
\description{
truncated factorial to avoid stack overflow for huge values
}
\examples{
truncfactorial(10)
}
\keyword{methods}
\keyword{models}
