% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.cpen.R
\name{estimate.gamma.cpen}
\alias{estimate.gamma.cpen}
\title{Estimate gamma cpen}
\usage{
estimate.gamma.cpen(
  formula,
  data,
  r = 1/1000,
  logn = log(1000),
  relat = c("cos", "sigmoid", "tanh", "atan", "sin", "erf")
)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{r}{r}

\item{logn}{logn}

\item{relat}{relations}
}
\description{
Estimate gamma cpen
}
