% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.gamma.cpen.R
\name{estimate.gamma.cpen_2}
\alias{estimate.gamma.cpen_2}
\title{Estimate gamma cpen 2}
\usage{
estimate.gamma.cpen_2(
  formula,
  data,
  r = 1/223,
  logn = log(223),
  relat = c("to23", "expi", "logi", "to35", "sini", "troot", "sigmoid")
)
}
\arguments{
\item{formula}{formula}

\item{data}{dataset}

\item{r}{r}

\item{logn}{logn}

\item{relat}{relations}
}
\description{
Estimate gamma cpen 2
}
