\name{m}
\alias{m}
\title{product function used in the deep regression context}
\description{product function used in the deep regression context}
\usage{m(a,b)}
\arguments{
\item{a}{the first argument}
\item{b}{the second argument}
}
\value{
\item{m(a,b)}{product of the arguments a*b}
}
\examples{EMJMCMC:::m(10,2)}
\keyword{methods}% use one of  RShowDoc("KEYWORDS")
\keyword{models}% __ONLY ONE__ keyword per line
