% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelize.R
\name{parallelize}
\alias{parallelize}
\title{An example of user defined parallelization (cluster based) function
for within an MJMCMC chain calculations (mclapply or lapply are used by
default depending on specification and OS).}
\usage{
parallelize(X, FUN)
}
\arguments{
\item{X}{a vector (atomic or list) or an expressions vector. Other objects
(including classed objects) will be coerced by as.list}

\item{FUN}{the function to be applied to each element of X or v, or in
parallel to X}
}
\value{
\code{parallelize(X,FUN)}, a list of the same length as X and named by X
}
\description{
An example of user defined parallelization (cluster based) function
for within an MJMCMC chain calculations (mclapply or lapply are used by
default depending on specification and OS).
}
\details{
Only allowed when working with big.memory based hash table within
MJMCMC (see runemjmcmc for more details)
}
\seealso{
parLapply clusterMap mclapply lapply
}
\keyword{methods}
\keyword{models}
