\name{sigmoid}
\alias{sigmoid}
\title{sigmoid activation function}
\usage{sigmoid(x)}
\arguments{
\item{x}{a real number}
}
\value{
\item{sigmoid(x)}{sigmoid value}
}
\examples{EMJMCMC:::sigmoid(10)}
\keyword{methods}% use one of  RShowDoc("KEYWORDS")
\keyword{models}% __ONLY ONE__ keyword per line
